requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.1
django-autocomplete-light==3.3.2
django-bootstrap3==11.0.0
django-cors-headers==2.4.0
django-filter==2.0.0
django-relativedelta==1.0.4
django-selectable==1.1.0
djangorestframework==3.8.2
idna==2.7
mysqlclient==1.3.13
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyflakes==2.0.0
PyJWT==1.6.4
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
scipy==1.1.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
